+++ Fogger/src/main/java/dotblueshoes/fogger/config/FogDefinition.java
@@ -2,18 +2,16 @@ package dotblueshoes.fogger.config;
 
 public class FogDefinition {
     public String name;
-    public float
-        fogMinIntensity,
-        fogMaxIntensity;
+    public float fogStartPoint, fogEndPoint;
 
-    public FogDefinition(String name, float fogMinIntensity, float fogMaxIntensity) {
-        this.fogMinIntensity = fogMinIntensity;
-        this.fogMaxIntensity = fogMaxIntensity;
+    public FogDefinition(String name, float fogStartPoint, float fogEndPoint) {
+        this.fogStartPoint = fogStartPoint;
+        this.fogEndPoint = fogEndPoint;
         this.name = name;
     }
 
     public String toString() {
-        return name + " " + Float.toString(fogMinIntensity) + " " + Float.toString(fogMaxIntensity);
+        return name + " " + Float.toString(fogStartPoint) + " " + Float.toString(fogEndPoint);
     }
 
     public static FogDefinition parseString(String data) {