+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.1] - 2021-03-06
+
+## Added
+- Few new default config values.
+- No restricitions to y-level order.
+
+## Changed
+- Few default config values.
+
+## Removed
+- Mistakenly left code that changed how list 3rd value behaved.
+
 ## [1.1.0] - 2021-03-02
 
 ### Added